Update authorized-accounts to inherit from superclass; Update more imports
