+++ app/models/authorized-citation-account.ts
@@ -1,22 +1,11 @@
-import { AsyncBelongsTo, attr, belongsTo } from '@ember-data/model';
+import { AsyncBelongsTo, belongsTo } from '@ember-data/model';
 
 import ExternalCitationServiceModel from './external-citation-service';
-import { AddonCredentialFields } from './authorized-storage-account';
-import UserReferenceModel from './user-reference';
-import OsfModel from './osf-model';
+import AuthorizedAccountModel from './authorized-account';
 
-export default class AuthorizedCitationAccountModel extends OsfModel {
-    @attr('fixstring') displayName!: string;
-    @attr('fixstringarray') scopes!: string[];
-    @attr('object') credentials?: AddonCredentialFields; // write-only
-    @attr('fixstring') readonly authUrl!: string; // Only returned when POSTing to /authorized-citation-accounts
-    @attr('boolean') readonly credentialsAvailable!: boolean;
-
+export default class AuthorizedCitationAccountModel extends AuthorizedAccountModel {
     @belongsTo('external-citation-service')
     citationService!: AsyncBelongsTo<ExternalCitationServiceModel> & ExternalCitationServiceModel;
-
-    @belongsTo('user-reference', { inverse: 'authorizedCitationAccounts' })
-    configuringUser!: AsyncBelongsTo<UserReferenceModel> & UserReferenceModel;
 }
 
 declare module 'ember-data/types/registries/model' {
