+++ lib/osf-components/addon/components/addons-service/user-addons-manager/component.ts
@@ -13,7 +13,8 @@ import Toast from 'ember-toastr/services/toast';
 import UserReferenceModel from 'ember-osf-web/models/user-reference';
 import Provider, {AllProviderTypes, AllAuthorizedAccountTypes} from 'ember-osf-web/packages/addons-service/provider';
 import CurrentUserService from 'ember-osf-web/services/current-user';
-import AuthorizedStorageAccountModel, { AddonCredentialFields } from 'ember-osf-web/models/authorized-storage-account';
+import { AddonCredentialFields } from 'ember-osf-web/models/authorized-account';
+import AuthorizedStorageAccountModel from 'ember-osf-web/models/authorized-storage-account';
 import AuthorizedCitationAccountModel from 'ember-osf-web/models/authorized-citation-account';
 import AuthorizedComputingAccount from 'ember-osf-web/models/authorized-computing-account';
 import UserModel from 'ember-osf-web/models/user';
